[SWM-144] refactor: 에러 처리 변경에 따른 테스트 코드 에러 메시지 검증 제거
