+++ src/test/java/com/climbx/climbx/auth/AuthServiceTest.java
@@ -1,7 +1,6 @@
 package com.climbx.climbx.auth;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.verify;
 
@@ -86,18 +85,6 @@ class AuthServiceTest {
         assertThat(response.provider()).isEqualTo("GOOGLE");
         assertThat(response.issuedAt()).isNotNull();
         assertThat(response.expiresAt()).isNotNull();
-    }
-
-    @Test
-    @DisplayName("유효하지 않은 user-id로 사용자 정보 조회 시 예외가 발생한다")
-    void shouldThrowExceptionForInvalidToken() {
-        // given
-        Long invalidUserId = 123456789L;
-
-        // when & then
-        assertThatThrownBy(() -> authService.getCurrentUserInfo(invalidUserId))
-            .isInstanceOf(UserUnauthorizedException.class)
-            .hasMessage("Unauthorized user");
     }
 
     @Test
